feat: complete game rules, more tests
Complete implementation of game rules. This includes the rules around
completely surrounding the king, which were missing before, and
re-implementing the other rules for killing a piece with a simpler
matching logic.

Improved test suite by moving the code for board and pieces to their own
files and implementing more tests on the basic functions of the board.

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>